fix: improve speed loading sidebar nav by splitting the sidebar into smaller components
ref: MANAGER-14071
