+++ packages/manager/apps/container/src/container/nav-reshuffle/sidebar/sidebar.type.ts
@@ -1,0 +1,12 @@
+export interface ServicesCount {
+    total: number;
+    serviceTypes: Record<string, number>;
+    errors?: Array<ServicesCountError>;
+  }
+
+  interface ServicesCountError {
+    url: string;
+    status: number;
+    message: string;
+  }
+
